merged: changes from linuxport revs r22314-r22365
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/pvr-testing/XBMC@22367 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/dvdplayer/DVDSubtitles/DVDSubtitleParserVplayer.cpp
@@ -48,7 +48,7 @@ bool CDVDSubtitleParserVplayer::Open(CDVDStreamInfo &hints)
   // Vplayer subtitles don't have StopTime, so we use following subtitle's StartTime
   // for that, unless gap was more than 4 seconds. Then we set subtitle duration
   // for 4 seconds, to not have text hanging around in silent scenes...
-  int iDefaultDuration = 4 * m_framerate;
+  int iDefaultDuration = 4 * (int)m_framerate;
 
   char line[1024];
